Removed references to obsolete DisplayFont
According to NEWS, the DisplayFont option was removed in version 0.62.0 and
replaced by the WINGs system font.  However, it was still referenced in a large
number of style files, as well as the utils getstyle, setstyle, convertfonts,
and the manpages for getstyle and setstyle.  This patch removes all of these
references.

In addition, it updates the Czech, Slovak, and Russian translations of these
manpages to match the English versions.  Note that I relied on Google translate, and it appears I may have re-encoded some characters when I saved the Czech
and Slovak files -- if you speak these languages, please take a look!

A few minor spelling errors were corrected in the English manpages.